add HoverArea to detect mouse hover on slide panels
We only care about enter and leave, not the other MouseArea features.
Of course there is the usual trouble that because this is on top,
it blocks hover events from going through to any other MouseAreas
underneath; whereas if we put it underneath, then if the mouse
"enters" a launcher button or something, it will "exit" from
HoverArea.  So, nothing inside the panel is allowed to use hoverable
MouseAreas for now.