add -s --screen option; move the window to that screen
There is preliminary support for using this option multiple times,
or iterating all available screens; but it seems we need to use the C++
API to manage the compositor then, instead of declaring Screen in main.qml.
For now, it's only useful to give this option once, and it's only useful
when running Grefsen under another windowing system like X11.